Vivo: Support Android encryption
Also reserve final 16kb of /data for encryption footer
Force recovery to reserve the final 16kb of the data partition
so that the encryption footer for phone encryption can be
stored there. Unlike newer devices we lack a dedicated partition
for the encryption footer so it gets stored at the end of /data.

Change-Id: I07c52f761da50dcd1b02b57dc8350fad691a42c4